+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.dto;
 
 import java.time.LocalDate;
-import roomescape.domain_entity.Reservation;
+import roomescape.entity.Reservation;
 
 public record ReservationResponse(
         Long id, String name, LocalDate date, ReservationTimeResponse time, ThemeResponse theme
