+++ src/main/java/roomescape/entity/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain_entity;
+package roomescape.entity;
 
 import java.util.Objects;
