+++ src/main/java/roomescape/service/ThemeService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Component;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
-import roomescape.domain_entity.Theme;
+import roomescape.entity.Theme;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
