+++ src/test/java/roomescape/entity/ReservationTest.java
@@ -1,9 +1,8 @@
-package roomescape.domain_entity;
+package roomescape.entity;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
