+++ src/test/java/roomescape/entity/ReservationTimeTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain_entity;
+package roomescape.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
